--- src/features/gathering/index.ts
@@ -1,2 +1,0 @@
-export * from './model/options';
-export * from './model/types';
